# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse API",
    "description": "AI-powered research & synthesis platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/research": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Start Research",
        "description": "Persist a new research job and hand it off to the worker.\n\nThe row is committed before the task is enqueued so the worker (which loads the job by id) is guaranteed to see it. If the enqueue fails the row is left in `pending`; a later sweep can either retry it or mark it as failed.",
        "operationId": "start_research_api_research_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Research",
        "description": "Paginated list of the caller's research jobs, newest first.",
        "operationId": "list_research_api_research_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/preview": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Preview Research",
        "description": "Run only Scout's decompose step and return the proposed sub-questions.\n\nNo job row is created and no work is queued. This is intentionally a\nsynchronous preview: the caller gets sub-questions back immediately so they\ncan review and drop them before committing to a full research run.",
        "operationId": "preview_research_api_research_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/research/{job_id}/report": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_research__job_id__report_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiedReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/{job_id}": {
      "delete": {
        "tags": [
          "research"
        ],
        "summary": "Delete Research",
        "description": "Delete one of the caller's research jobs and everything derived from it.\n\nFollow-up children are not deleted \u2014 the parent/child link is dropped and they\nremain as standalone research jobs.",
        "operationId": "delete_research_api_research__job_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/{job_id}/export/markdown": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Export Markdown",
        "operationId": "export_markdown_api_research__job_id__export_markdown_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/{job_id}/export/pdf": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Export Pdf",
        "operationId": "export_pdf_api_research__job_id__export_pdf_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaimFlag": {
        "properties": {
          "claim_id": {
            "type": "string",
            "title": "Claim Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "supporting_source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Source Ids"
          }
        },
        "type": "object",
        "required": [
          "claim_id",
          "section_id",
          "verdict",
          "rationale",
          "supporting_source_ids"
        ],
        "title": "ClaimFlag"
      },
      "Contradiction": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "type": "object",
        "required": [
          "description",
          "source_ids"
        ],
        "title": "Contradiction"
      },
      "CriticAnnotations": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Report Id"
          },
          "section_confidence": {
            "items": {
              "$ref": "#/components/schemas/SectionConfidence"
            },
            "type": "array",
            "title": "Section Confidence"
          },
          "claim_flags": {
            "items": {
              "$ref": "#/components/schemas/ClaimFlag"
            },
            "type": "array",
            "title": "Claim Flags"
          },
          "overall_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Confidence"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_id",
          "section_confidence",
          "claim_flags",
          "overall_confidence",
          "model",
          "generated_at"
        ],
        "title": "CriticAnnotations"
      },
      "Depth": {
        "type": "string",
        "enum": [
          "shallow",
          "standard",
          "deep"
        ],
        "title": "Depth"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "JobListResponse",
        "description": "Paginated envelope for the history list."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "scouting",
          "synthesizing",
          "critiquing",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0.0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count",
            "default": 0
          },
          "overall_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Confidence"
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Job Id"
          },
          "parent_topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Topic"
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic",
          "status",
          "created_at"
        ],
        "title": "JobSummary",
        "description": "Compact job descriptor for the history list (`GET /api/research`).\n\nCarries only what the library list renders, plus the follow-up parent edge so a child job\ncan show a \"Follow-up of \u2026\" link. `source_count` and `overall_confidence` are derived by\njoin at query time; `overall_confidence` is null until the job completes and Critic writes its\nannotations."
      },
      "PreviewResponse": {
        "properties": {
          "sub_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sub Questions"
          }
        },
        "type": "object",
        "required": [
          "sub_questions"
        ],
        "title": "PreviewResponse",
        "description": "Response body for POST /api/research/preview."
      },
      "ReportSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "heading": {
            "type": "string",
            "title": "Heading"
          },
          "body_md": {
            "type": "string",
            "title": "Body Md"
          },
          "cited_source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cited Source Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "heading",
          "body_md"
        ],
        "title": "ReportSection"
      },
      "ResearchJob": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "depth": {
            "$ref": "#/components/schemas/Depth",
            "default": "standard"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Models"
          },
          "sub_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Questions"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0.0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic"
        ],
        "title": "ResearchJob",
        "description": "Job descriptor returned to the client on creation and status queries."
      },
      "ResearchRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Topic"
          },
          "language": {
            "type": "string",
            "maxLength": 8,
            "minLength": 2,
            "title": "Language",
            "default": "en"
          },
          "depth": {
            "$ref": "#/components/schemas/Depth",
            "default": "standard"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Models"
          },
          "sub_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Questions"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "models"
        ],
        "title": "ResearchRequest",
        "description": "Inbound request body for POST /api/research."
      },
      "ScribeReport": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary_md": {
            "type": "string",
            "title": "Summary Md"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "contradictions": {
            "items": {
              "$ref": "#/components/schemas/Contradiction"
            },
            "type": "array",
            "title": "Contradictions"
          },
          "follow_ups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Follow Ups"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "topic",
          "title",
          "summary_md",
          "sections",
          "sources",
          "contradictions",
          "follow_ups",
          "generated_at",
          "model"
        ],
        "title": "ScribeReport"
      },
      "SectionConfidence": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "score",
          "reasoning"
        ],
        "title": "SectionConfidence"
      },
      "Source": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "credibility": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Credibility"
          },
          "relevance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "credibility",
          "relevance",
          "snippet"
        ],
        "title": "Source",
        "description": "A single source gathered by Scout, referenced by Scribe and Critic."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "supported",
          "partially_supported",
          "unsupported",
          "contradicted"
        ],
        "title": "Verdict"
      },
      "VerifiedReport": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ResearchJob"
          },
          "report": {
            "$ref": "#/components/schemas/ScribeReport"
          },
          "annotations": {
            "$ref": "#/components/schemas/CriticAnnotations"
          }
        },
        "type": "object",
        "required": [
          "job",
          "report",
          "annotations"
        ],
        "title": "VerifiedReport",
        "description": "Full response returned to the frontend once a job has completed."
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "synapse_auth"
      }
    }
  }
}